Update tests, move most functionality to exports.
